#F-89 Fixed java.nio.file.NoSuchFileException: tempConverterFile.bin

+++ Communication/src/main/java/com/ak/comm/file/FileReadingService.java
@@ -82,7 +82,7 @@ final class FileReadingService<RESPONSE, REQUEST, EV extends Enum<EV> & Variable
               process(byteBuffer).forEach(s::onNext);
             });
 
-            if (processed) {
+            if (processed && Files.exists(tempConverterFile)) {
               Files.copy(tempConverterFile, convertedFile, LinkOption.NOFOLLOW_LINKS, StandardCopyOption.REPLACE_EXISTING);
             }
           }
